+++ R/loadDetectionData.R
@@ -1,40 +1,41 @@
 #' @title Load Detection Data
-#' 
-#' @description Loads and formats detection data into a common format for 
+#'
+#' @description Loads and formats detection data into a common format for
 #'   use in other PAMscapes functions
-#'   
+#'
 #' @param x dataframe or path to CSV file containing detection data
 #' @param source source of the detection data, choices other than "csv"
 #'   just specify specific formatting options
 #' @param columnMap a list or data.frame specifying how to map the input
 #'   column names to the required standard names of "UTC", "end", and "species".
-#'   If a list, must be a named list where the names are the existing 
-#'   column names and the values are the standardized names, e.g. 
-#'   \code{list('start'='UTC', 'SpeciesName'='species')}. If a data.frame,
+#'   If a list, must be a named list where the names are the standardized
+#'   column names and the values are the existing names, e.g.
+#'   \code{list('UTC'='start', 'species'='SpeciesName')}. If a data.frame,
 #'   must have columns "old" with the existing column names and "new" with
 #'   the standardized name to change it to. All columns successfully changed
 #'   will be kept with the output
 #' @param detectionType one of "auto", "presence", or "detection" specifying
 #'   the type of detection in the data. "presence" means hourly or daily presence
-#'   style of detections - the duration of the detection is used for the time 
+#'   style of detections - the duration of the detection is used for the time
 #'   unit (e.g. hourly presence might have "UTC" value 2020-01-01 12:00:00 and
 #'   "end" value 2020-01-01 13:00:00 for a detection). "detection" means the data
 #'   refer to specific detections or bouts of detections rather than just presence.
-#'   "auto" means that the type of detection will be inferred from the start and 
-#'   end time of each detection - any detections with a duration of exactly one 
+#'   "auto" means that the type of detection will be inferred from the start and
+#'   end time of each detection - any detections with a duration of exactly one
 #'   hour or exactly one day will be marked as "presence", any other duration
 #'   will be marked as "detection"
 #' @param presenceDuration if \code{detectionType='presence'}, the duration in
-#'   seconds, e.g. 86400 for daily presence
-#' @param dateFormat format string of dates, see \link{strptime}. Can be a 
+#'   seconds, e.g. 86400 for daily presence. Alternative can be a character
+#'   of the form "(NUMBER)(DURATION)" e.g "2hour" or "1day"
+#' @param dateFormat format string of dates, see \link{strptime}. Can be a
 #'   vector of multiple formats
 #' @param tz time zone of input data
-#' @param wide logical flag indicating whether the input data has species 
+#' @param wide logical flag indicating whether the input data has species
 #'   detection information in wide (instead of long) format. If \code{TRUE},
 #'   then this means that there are multiple columns representing multiple
 #'   kinds of detections, e.g. one column for each different species present.
 #'   If \code{FALSE}, then there is a single column that indicates what kind
-#'   of detection it is. 
+#'   of detection it is.
 #' @param speciesCols only used if \code{wide=TRUE}, the names of the columns
 #'   containing the different types of detections
 #' @param detectedValues only used if \code{wide=TRUE}, the values in each
@@ -44,21 +45,21 @@
 #'   so the string \code{"1"} must be used instead of the numeric \code{1}
 #' @param extraCols (optional) any additional columns to keep with the output
 #' @param \dots additional arguments used for certain \code{source} values
-#' 
+#'
 #' @author Taiki Sakai \email{taiki.sakai@@noaa.gov}
-#' 
+#'
 #' @return a dataframe with columns UTC, end, species, and detectionType, where
-#'   each row represents a single detection event. May have additional columns 
+#'   each row represents a single detection event. May have additional columns
 #'   depending on other parameters
-#'   
+#'
 #' @export
-#' 
+#'
 #' @importFrom tidyr pivot_longer
 #' @importFrom lubridate parse_date_time
 #' @importFrom utils read.csv
 #'
 loadDetectionData <- function(x,
-                              source=c('csv', 'makara'),
+                              source=c('makara', 'csv'),
                               columnMap=NULL,
                               detectionType=c('auto', 'presence', 'detection'),
                               presenceDuration=NULL,
@@ -68,14 +69,18 @@ loadDetectionData <- function(x,
                                            '%Y/%m/%d %H:%M:%S',
                                            '%m/%d/%Y %H:%M:%S'),
                               tz='UTC',
-                              wide=FALSE, 
+                              wide=FALSE,
                               speciesCols=NULL,
                               detectedValues=NULL,
                               extraCols=NULL,
                               ...) {
+    # Allow for '1day' style description of presence
+    if(!is.null(presenceDuration) && is.character(presenceDuration)) {
+        presenceDuration <- as.numeric(unitToPeriod(presenceDuration))
+    }
     if(is.character(x) && length(x) > 1) {
         result <- bind_rows(lapply(x, function(d) {
-            loadDetectionData(d, 
+            loadDetectionData(d,
                               source=source,
                               columnMap=columnMap,
                               detectionType='detection',
@@ -129,6 +134,10 @@ loadDetectionData <- function(x,
     }
     if(source == 'makara') {
         columnMap <- getColMaps('makara')
+        if(!any(unlist(columnMap) %in% names(result))) {
+            warning('Input does not appear to be Makara data, change "source"')
+            return(NULL)
+        }
         result$deployment <- detFileToCode(x)
         extraCols <- c(extraCols, 'call', 'deployment')
         if(is.null(detectedValues)) {
@@ -149,10 +158,10 @@ loadDetectionData <- function(x,
         warning('Time zone ', tz, ' is invalid, must be present in "OlsonNames()"')
         return(NULL)
     }
-    result$UTC <- parse_date_time(result$UTC, 
-                                  orders=dateFormat, 
+    result$UTC <- parse_date_time(result$UTC,
+                                  orders=dateFormat,
                                   truncated = 3,
-                                  tz=tz, 
+                                  tz=tz,
                                   quiet=TRUE,
                                   exact=TRUE)
     if(tz != 'UTC') {
@@ -167,10 +176,10 @@ loadDetectionData <- function(x,
         warning(sum(naDate), ' dates could not be parsed correctly')
     }
     if('end' %in% names(result)) {
-        result$end <- parse_date_time(result$end, 
-                                      orders=dateFormat, 
+        result$end <- parse_date_time(result$end,
+                                      orders=dateFormat,
                                       truncated = 3,
-                                      tz=tz, 
+                                      tz=tz,
                                       quiet=TRUE,
                                       exact=TRUE)
     }
@@ -179,7 +188,7 @@ loadDetectionData <- function(x,
         result$end <- result$UTC + result$duration
         extraCols <- c(extraCols, 'duration')
     }
-    
+
     detectionType <- match.arg(detectionType)
     switch(detectionType,
            'auto' = {
@@ -214,8 +223,8 @@ loadDetectionData <- function(x,
             warning('Must specify which columns contain detection data with',
                     ' "speciesCols"')
         }
-        result <- pivot_longer(result, 
-                               cols=all_of(speciesCols), 
+        result <- pivot_longer(result,
+                               cols=all_of(speciesCols),
                                values_transform=as.character,
                                names_to='species',
                                values_to='detectedFlag')
@@ -232,28 +241,28 @@ loadDetectionData <- function(x,
 
 getColMaps <- function(which=NULL) {
     maps <- list(
-        'makara' = list('detection_start_datetime' = 'UTC',
-                        'detection_end_datetime' = 'end',
-                        'detection_sound_source_code' = 'species',
-                        'detection_call_type_code' = 'call',
-                        'detection_type_code' = 'detectedFlag',
-                        'detection_latitude' = 'Latitude',
-                        'detection_longitude' = 'Longitude'
+        'makara' = list('UTC' = 'detection_start_datetime',
+                        'end' = 'detection_end_datetime',
+                        'species' = 'detection_sound_source_code',
+                        'call' = 'detection_call_type_code',
+                        'detectedFlag' = 'detection_type_code',
+                        'Latitude' = 'detection_latitude',
+                        'Longitude' = 'detection_longitude'
         ),
-        'gen' = list('StartDate' = 'UTC',
-                     'PROJECT_DESCRIPTION' = 'project',
-                     'SITE_NAME' = 'site',
-                     'LATITUDE_DDG_DEPLOYMENT' = 'Latitude',
-                     'LONGITUDE_DDG_DEPLOYMENT' = 'Longitude'
+        'gen' = list('UTC' = 'StartDate',
+                     'project' = 'PROJECT_DESCRIPTION',
+                     'site' = 'SITE_NAME',
+                     'Latitude' = 'LATITUDE_DDG_DEPLOYMENT',
+                     'Longitude' = 'LONGITUDE_DDG_DEPLOYMENT'
         ),
-        'standard' = list('utc' = 'UTC',
+        'standard' = list('UTC'= 'utc' ,
                           'end' = 'end',
-                          'lat' = 'Latitude',
-                          'lon' = 'Longitude',
-                          'long' = 'Longitude',
-                          'latitude' = 'Latitude',
-                          'longitude' = 'Longitude',
-                          'start' = 'UTC',
+                          'Latitude' = 'lat',
+                          'Longitude' = 'lon',
+                          'Longitude' = 'long',
+                          'Latitude' = 'latitude',
+                          'Longitude' = 'longitude',
+                          'UTC' = 'start',
                           'duration' = 'duration',
                           'species' = 'species'
         )
@@ -299,9 +308,16 @@ inferDetType <- function(start, end, verbose=TRUE) {
 renameToMap <- function(names, map) {
     map <- checkMap(map)
     lowNames <- tolower(names)
-    map$old <- tolower(map$old)
-    for(i in which(lowNames %in% map$old)) {
-        names[i] <- map[['new']][map$old == lowNames[i]]
+    lowOld <- tolower(map$old)
+    # # in case old and new seem obviously swapped, reverse and redo
+    # if(!any(lowOld %in% lowNames) &&
+    #    any(tolower(map$new) %in% lowNames)) {
+    #     names(map) <- rev(names(map))
+    #     return(renameToMap(names, map))
+    # }
+
+    for(i in which(lowNames %in% lowOld)) {
+        names[i] <- map[['new']][lowOld == lowNames[i]]
     }
     names
 }
@@ -328,11 +344,11 @@ checkMap <- function(map) {
     }
     if(is.list(map) &&
        !is.data.frame(map)) {
-        map <- data.frame(old=names(map), new=unlist(map, use.names=FALSE))
+        map <- data.frame(new=names(map), old=unlist(map, use.names=FALSE))
     }
     if(is.character(map) &&
        !is.null(names(map))) {
-        map <- data.frame(old=names(map), new=unname(map))
+        map <- data.frame(new=names(map), old=unname(map))
     }
     if(!is.data.frame(map) ||
        !all(c('old', 'new') %in% names(map))) {